+++ src/main/java/com/sochoeun/service/impl/CategoryServiceImpl.java
@@ -1,6 +1,9 @@
 package com.sochoeun.service.impl;
 
-import com.sochoeun.exception.NotFoundException;
+import com.sochoeun.exception.CategoryException;
+import com.sochoeun.exception.ResourceNotFoundException;
+import com.sochoeun.model.Article;
+import com.sochoeun.repository.ArticleRepository;
 import com.sochoeun.repository.CategoryRepository;
 import com.sochoeun.model.Category;
 import com.sochoeun.service.CategoryService;
@@ -13,6 +16,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final ArticleRepository articleRepository;
     @Override
     public Category createCategory(Category request) {
         return categoryRepository.save(request);
@@ -25,7 +29,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category getCategory(Integer categoryId) {
-        return categoryRepository.findById(categoryId).orElseThrow(()->new NotFoundException("Category",categoryId));
+        return categoryRepository.findById(categoryId).orElseThrow(()->new ResourceNotFoundException("Category ID: %s not found.".formatted(categoryId)));
 
     }
 
@@ -41,7 +45,12 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategory(Integer categoryId) {
-        getCategory(categoryId); // get not found -> response: not found
+        getCategory(categoryId);// get not found -> response: not found
+        List<Article> allByCategoryId = articleRepository.findAllByCategory_Id(categoryId);
+
+        if (!allByCategoryId.isEmpty()){
+            throw new CategoryException(categoryId);
+        }
         categoryRepository.deleteById(categoryId);
     }
 }
